Implements a new phi-placement function
Previous verison was subtly flawed for corner cases.
This version places a dummy/placeholder-phi in the nodes dominance frontier
for each variable written. This includes any phi-nodes being defined. Due
to this an iterative approach is taken, it continues untill the processs
stabilizes. Once stable, ssa_ir phi nodes are inserted and by that time
the exact source for definitions in the nodes are determined.