Add `state_variable_slots` property to contract.py
I am using this in my fork, and thought it would be a nice and useful enough addition on its own to make opening a PR worthwhile.
I will be using it in a new upgradeability check I am working on, which will detect all of the differences between V1 and V2 of an implementation contract, for informational purposes and to feed into the differential fuzzing test framework I aim to design. Knowing the storage slot for each state variable will be useful when using cheat codes to read from a contract's storage and compare the values between two versions.